+++ CHANGES.md
@@ -1,9 +1,13 @@
 CHANGES
 ========
 
+## 1.1.12
+
+* Feature: Remove Comment Button
+
 ## 1.1.11
 
-* Fix: omit comment form from page_list (crowi-plus Enhanced Layout)
+* Fix: Omit Comment form from page_list (crowi-plus Enhanced Layout)
 * Fix: .search-box is broken on sm/xs screen
 
 ## 1.1.10